+++ R/models.R
@@ -13,7 +13,7 @@
 #' controls <- list(
 #'   states = 2,
 #'   sdds   = "t(df = Inf)",
-#'   data   = list(file        = "inst/extdata/dax.csv",
+#'   data   = list(file        = system.file("extdata", "dax.csv", package = "fHMM"),
 #'                 date_column = "Date",
 #'                 data_column = "Close",
 #'                 logreturns  = TRUE,
@@ -22,7 +22,6 @@
 #'   fit    = list("runs" = 100)
 #'   )
 #' }
-#' Set `file = "extdata/dax.csv"` to reproduce this code.
 #'
 #' @format An object of class \code{fHMM_model}.
 #'
@@ -45,7 +44,7 @@
 #' controls <- list(
 #'   states = 3,
 #'   sdds   = "t",
-#'   data   = list(file        = "inst/extdata/dax.csv",
+#'   data   = list(file        = system.file("extdata", "dax.csv", package = "fHMM"),
 #'                 date_column = "Date",
 #'                 data_column = "Close",
 #'                 logreturns  = TRUE,
@@ -54,7 +53,6 @@
 #'   fit    = list("runs" = 100)
 #'   )
 #' }
-#' Set `file = "extdata/dax.csv"` to reproduce this code.
 #'
 #' @format An object of class \code{fHMM_model}.
 #'
@@ -81,17 +79,48 @@
 #'   states    = c(2,2),
 #'   sdds      = c("t(df = 1)", "t(df = 1)"),
 #'   period    = "m",
-#'   data      = list(file = c("inst/extdata/dax.csv", "inst/extdata/vw.csv"),
+#'   data      = list(file = c(system.file("extdata", "dax.csv", package = "fHMM"),
+#'                             system.file("extdata", "vw.csv", package = "fHMM")),
 #'                    from = "2015-01-01",
 #'                    to = "2020-01-01",
-#'                    logreturns = c(TRUE,TRUE)),
-#'   fit       = list("runs" = 100)
+#'                    logreturns = c(TRUE,TRUE))
 #' )
 #' }
-#' Set `file = c("extdata/dax.csv", "extdata/vw.csv")` to reproduce this code.
 #'
 #' @format An object of class \code{fHMM_model}.
 #'
 #' @keywords
 #' model
 "dax_vw_model"
+
+#' Simulated 2-state HMM
+#'
+#' @description
+#' A pre-computed 2-state HMM with state-dependent gamma distributions with means
+#' fixed to \code{0.5} and \code{2} on \code{500} simulated observations.
+#'
+#' @usage data(sim_model_2gamma)
+#'
+#' @details
+#' The model was estimated via:
+#' \preformatted{
+#' controls <- list(
+#'   states  = 2,
+#'   sdds    = "gamma(mu = 1|2)",
+#'   horizon = 200,
+#'   fit     = list(runs = 50)
+#' )
+#' controls <- set_controls(controls)
+#' pars <- fHMM_parameters(
+#'   controls = controls, Gamma = matrix(c(0.9,0.2,0.1,0.8), nrow = 2),
+#'   sigmas = c(0.5,1)
+#' )
+#' data <- prepare_data(controls, true_parameters = pars, seed = 1)
+#' sim_model_2gamma <- fit_model(data, seed = 1, verbose = FALSE)
+#' }
+#'
+#' @format An object of class \code{fHMM_model}.
+#'
+#' @keywords
+#' model
+"sim_model_2gamma"