+++ NEWS.md
@@ -1,3 +1,10 @@
+# fHMM 1.0.1
+
+* In the vignette 'Controls', in the section about example specifications for `controls`, corrected `sdds = "gamma(mu = -1|1)"` to `sdds = "gamma(mu = 0.5|2)"` because mean of the Gamma distribution must be positive.
+* Added `digits` argument to `print.fHMM_predict()`.
+* Fixed bug in `reorder_states()` that allowed for misspecification of `state_order`.
+* Added option to `fit_model()` to initialize at the estimates of another model (#73).
+
 # fHMM 1.0.0
 
 * Enhanced the package by S3 classes.